Updated most API endpoints to require authentication

+++ api/controllers/entries.js
@@ -5,26 +5,36 @@ const pool = createDbConnection()
 
 export const createJournalEntry = (req, res) => {
     const newJournalEntry = req.body
-    const insertStatementContent = `'${uuidv4()}', '${newJournalEntry.associated_journal}', '${newJournalEntry.title}', '${newJournalEntry.content}', CURRENT_TIMESTAMP`.toString()
-    pool.query(`INSERT INTO journal_entries (journal_entry_id, associated_journal, title, content, date_created) VALUES (${insertStatementContent});`, (error, results) => {
-        if (error) {
-            res.status(400).json(error.message)
-        } else {
-            res.status(201).send('Journal Entry Added')
-        }
-    })
+    const { id } = req.user
+    try {
+        const insertStatementContent = `'${uuidv4()}', '${newJournalEntry.associated_journal}', '${newJournalEntry.title}', '${newJournalEntry.content}', CURRENT_TIMESTAMP, '${id}'`.toString()
+        pool.query(`INSERT INTO journal_entries (journal_entry_id, associated_journal, title, content, date_created, owner) VALUES (${insertStatementContent});`, (error, results) => {
+            if (error) {
+                res.status(400).json(error.message)
+            } else {
+                res.status(201).send('Journal Entry Added')
+            }
+        })
+    } catch {
+        res.sendStatus(500)
+    }
 }
 
 export const getJournalEntry = (req, res) => {
     const { journalEntryId } = req.params;
-    pool.query(
-        `SELECT * FROM journal_entries WHERE journal_entry_id='${journalEntryId}';`, (error, results) => {
-        if (error) {
-            res.status(400).json(error.message)
-        } else {
-            res.status(200).json(results.rows)
-        }
-    })
+    const { id } = req.user
+    try {
+        pool.query(
+            `SELECT * FROM journal_entries WHERE journal_entry_id='${journalEntryId}' AND owner='${id}';`, (error, results) => {
+            if (error) {
+                res.status(400).json(error.message)
+            } else {
+                res.status(200).json(results.rows)
+            }
+        })
+    } catch {
+        res.sendStatus(500)
+    }
 }
 
 export const updateJournalEntry = (req, res) => {
@@ -44,11 +54,16 @@ export const updateJournalEntry = (req, res) => {
 
 export const deleteJournalEntry = (req, res) => {
     const { journalEntryId } = req.params;
-    pool.query(`DELETE FROM journal_entries WHERE journal_entry_id = '${journalEntryId}'`, (error, results) => {
-        if(error) {
-            res.status(400).json(error.message)
-        } else {
-            res.status(200).json('Success')
-        }
-    })
+    const { id } = req.user
+    try {
+        pool.query(`DELETE FROM journal_entries WHERE journal_entry_id = '${journalEntryId}' AND owner='${id}'`, (error, results) => {
+            if(error) {
+                res.status(400).json(error.message)
+            } else {
+                res.status(200).json('Success')
+            }
+        })
+    } catch {
+        res.sendStatus(500)
+    }
 }
